refactor: Move `SFTPServer` to separate helper module
